requirements.txt
aiofiles==24.1.0
aiogram==3.22.0        
aiohappyeyeballs==2.6.1
aiohttp==3.12.15       
aiosignal==1.4.0       
annotated-types==0.7.0 
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26     
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
frozenlist==1.7.0
gigachat
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.62
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
magic-filter==1.0.12
multidict==6.6.4
numpy>=2.0.0
openai==1.82.0
orjson==3.10.18
packaging==24.2
pandas>=2.3.3
pillow==11.3.0
xlsxwriter>=3.0.0
scikit-learn>=1.3.0
propcache==0.3.2
pydantic==2.11.5
pydantic_core==2.33.2
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.1.0
python-telegram-bot==22.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0